refactor(fast-usdc): finishDeploy
reconsider ExtraCorePowers: infer from permit
